requirements.txt
# Core web framework and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Numerical and scientific computing
numpy==1.24.3

# Deep learning and AI
torch==2.1.0
transformers==4.36.2

# Image processing
pillow==10.0.1

# Vector search and indexing
faiss-cpu==1.7.4

# Computer vision and media processing
opencv-python==4.8.1.78